move ldap definition to its own file for simplicity & context
